autogen-ethical-hacker/requirements.txt
autogen[all]
langchain
openai
groq
google-genai
python-nmap
redis
networkx
matplotlib
docker
fastapi
uvicorn
pydantic
streamlit
pyyaml
requests
